Split extensions into more files
