+++ GDEdit/GDEdit/Utilities/Functions/General/Extensions.cs
@@ -9,359 +9,6 @@ namespace GDEdit.Utilities.Functions.General
 {
     public static class Extensions
     {
-        public static int Find(this string s, string match)
-        {
-            for (int i = 0; i <= s.Length - match.Length; i++)
-            {
-                bool found = true;
-                for (int j = 0; j < match.Length && found; j++)
-                    if (s[i + j] != match[j])
-                        found = false;
-                if (found)
-                    return i;
-            }
-            return -1;
-        }
-        public static int Find(this string s, string match, int occurence)
-        {
-            int occurences = 0;
-            for (int i = 0; i <= s.Length - match.Length; i++)
-            {
-                bool found = true;
-                for (int j = 0; j < match.Length && found; j++)
-                    if (s[i + j] != match[j])
-                        found = false;
-                if (found)
-                {
-                    occurences++;
-                    if (occurences == occurence)
-                        return i;
-                }
-            }
-            return -1;
-        }
-        public static int Find(this string s, string match, int start, int end)
-        {
-            for (int i = start; i <= end - match.Length; i++)
-            {
-                bool found = true;
-                for (int j = 0; j < match.Length && found; j++)
-                    if (s[i + j] != match[j])
-                        found = false;
-                if (found)
-                    return i;
-            }
-            return -1;
-        }
-        public static int FindFromEnd(this string s, string match)
-        {
-            for (int i = s.Length - match.Length; i >= 0; i--)
-            {
-                bool found = true;
-                for (int j = 0; j < match.Length && found; j++)
-                    if (s[i + j] != match[j])
-                        found = false;
-                if (found)
-                    return i;
-            }
-            return -1;
-        }
-        public static int FindFromEnd(this string s, string match, int occurence)
-        {
-            int occurences = 0;
-            for (int i = s.Length - match.Length; i >= 0; i--)
-            {
-                bool found = true;
-                for (int j = 0; j < match.Length && found; j++)
-                    if (s[i + j] != match[j])
-                        found = false;
-                if (found)
-                {
-                    occurences++;
-                    if (occurences == occurence)
-                        return i;
-                }
-            }
-            return -1;
-        }
-        public static int FindFromEnd(this string s, string match, int start, int end)
-        {
-            for (int i = s.Length - match.Length; i >= 0; i--)
-            {
-                bool found = true;
-                for (int j = 0; j < match.Length && found; j++)
-                    if (s[i + j] != match[j])
-                        found = false;
-                if (found)
-                    return i;
-            }
-            return -1;
-        }
-        public static int GetLastNumber(this string s)
-        {
-            int n = 0;
-            bool isLastNumber = false;
-            for (int i = 0; i < s.Length; i++)
-            {
-                isLastNumber = int.TryParse(s.Substring(i), out n);
-                if (isLastNumber)
-                    return n;
-            }
-            throw new ArgumentException("The string has no number in the end.");
-        }
-        public static int[] FindAll(this string s, string match)
-        {
-            List<int> indices = new List<int>();
-            for (int i = 0; i <= s.Length - match.Length; i++)
-            {
-                string sub = s.Substring(i, match.Length);
-                if (sub == match)
-                    indices.Add(i);
-            }
-            return indices.ToArray();
-        }
-        public static int[] FindAll(this string s, string match, int start, int end)
-        {
-            List<int> indices = new List<int>();
-            for (int i = start; i <= end - match.Length; i++)
-            {
-                string sub = s.Substring(i, match.Length);
-                if (sub == match)
-                    indices.Add(i);
-            }
-            return indices.ToArray();
-        }
-        public static string FixBase64String(this string s)
-        {
-            int lastNumber = 0;
-            int lastInvalidCharacter = 0;
-            bool continueChecking = true;
-            while (continueChecking)
-            {
-                if (s[s.Length - lastNumber - 1].IsNumber())
-                    lastNumber++;
-                else if (!s[s.Length - lastNumber - 1].IsBase64Character())
-                    lastInvalidCharacter = ++lastNumber;
-                else
-                    continueChecking = false;
-            }
-            s = s.Substring(0, s.Length - lastInvalidCharacter);
-            while (s.Length % 4 != 0)
-                s += "=";
-            return s;
-        }
-        public static string Substring(this string s, string from, string to)
-        {
-            int startIndex = s.Find(from) + from.Length;
-            int endIndex = s.Find(to);
-            int length = endIndex - startIndex;
-            return s.Substring(startIndex, length);
-        }
-        public static string RemoveLastNumber(this string s)
-        {
-            string l = s;
-            for (int i = 0; i < s.Length; i++)
-            {
-                l = s.Substring(i);
-                if (int.TryParse(l, out int shit))
-                    return s.Substring(0, i);
-            }
-            return s;
-        }
-        public static string Replace(this string originalString, string stringToReplaceWith, int startIndex, int length)
-        {
-            string result = originalString;
-            result = result.Remove(startIndex, length);
-            result = result.Insert(startIndex, stringToReplaceWith);
-            return result;
-        }
-        public static string ReplaceWholeWord(this string originalString, string oldString, string newString)
-        {
-            for (int i = originalString.Length - oldString.Length; i >= 0; i--)
-            {
-                if (originalString.Substring(i, oldString.Length) == oldString)
-                    if ((i > 0 ? (!originalString[i - 1].IsLetterOrNumber()) : true) && (i < originalString.Length - oldString.Length ? (!originalString[i + oldString.Length].IsLetterOrNumber()) : true))
-                    {
-                        originalString = originalString.Replace(newString, i, oldString.Length);
-                        i -= oldString.Length;
-                    }
-            }
-            return originalString;
-        }
-        public static string[] Split(this string s, string separator)
-        {
-            int[] occurences = s.FindAll(separator);
-            string[] result = new string[occurences.Length + 1];
-            for (int i = 0; i < result.Length; i++)
-            {
-                int startIndex = i == 0 ? 0 : occurences[i - 1] + separator.Length;
-                int endIndex = i >= occurences.Length ? s.Length : occurences[i];
-                result[i] = s.Substring(startIndex, endIndex - startIndex);
-            }
-            return result;
-        }
-        public static bool Contains(this string s, params char[] c)
-        {
-            for (int i = 0; i < s.Length; i++)
-                for (int j = 0; j < c.Length; j++)
-                    if (s[i] == c[j])
-                        return true;
-            return false;
-        }
-        public static bool ContainsWholeWord(this string s, string match)
-        {
-            for (int i = s.Length - match.Length; i >= 0; i--)
-            {
-                if (s.Substring(i, match.Length) == match)
-                    if ((i > 0 ? (!s[i - 1].IsLetterOrNumber()) : true) && (i < s.Length - match.Length ? (!s[i + match.Length].IsLetterOrNumber()) : true))
-                        return true;
-            }
-            return false;
-        }
-        public static bool MatchesStringCaseFree(this string s, string match)
-        {
-            if (s.Length != match.Length)
-                return false;
-            for (int i = 0; i < s.Length; i++)
-            {
-                if (s[i].IsUpperCaseLetter())
-                    if (s[i] + 32 != match[i] && s[i] != match[i])
-                        return false;
-                else if (s[i].IsLowerCaseLetter())
-                    if (s[i] - 32 != match[i] && s[i] != match[i])
-                        return false;
-            }
-            return true;
-        }
-        public static bool IsBase64Character(this char c)
-        {
-            return IsNumber(c) || IsLetter(c) || c == '/' || c == '+' || c == '=';
-        }
-        public static bool IsLetter(this char c) => (c >= 'A' && c <= 'Z') || (c >= 'a' && c <= 'z');
-        public static bool IsLowerCaseLetter(this char c) => c >= 'a' && c <= 'z';
-        public static bool IsNumber(this char c) => c >= '0' && c <= '9';
-        public static bool IsUpperCaseLetter(this char c) => c >= 'A' && c <= 'Z';
-        public static bool IsLetterOrNumber(this char c) => IsLetter(c) || IsNumber(c);
-
-        public static int[] FindOccurences(this string[] a, string match)
-        {
-            if (a != null)
-            {
-                List<int> occurences = new List<int>();
-                for (int i = 0; i < a.Length; i++)
-                    if (a[i] == match)
-                        occurences.Add(i);
-                return occurences.ToArray();
-            }
-            else return new int[0];
-        }
-        public static string[] Replace(this string[] a, char oldChar, char newChar)
-        {
-            for (int i = 0; i < a.Length; i++)
-                a[i] = a[i].Replace(oldChar, newChar);
-            return a;
-        }
-        public static string[] Replace(this string[] a, string oldString, string newString)
-        {
-            for (int i = 0; i < a.Length; i++)
-                a[i] = a[i].Replace(oldString, newString);
-            return a;
-        }
-        public static string[] ReplaceWholeWord(this string[] a, string oldString, string newString)
-        {
-            for (int i = 0; i < a.Length; i++)
-                a[i] = a[i].ReplaceWholeWord(oldString, newString);
-            return a;
-        }
-        public static string[] RemoveEmptyElements(this string[] a)
-        {
-            List<string> result = new List<string>();
-            for (int i = 0; i < a.Length; i++)
-                if (a[i].Length > 0)
-                    result.Add(a[i]);
-            return result.ToArray();
-        }
-        public static bool ContainsAtLeast(this string[] a, string match)
-        {
-            if (a != null)
-            {
-                for (int i = 0; i < a.Length; i++)
-                    if (a[i].Contains(match))
-                        return true;
-                return false;
-            }
-            return false;
-        }
-        public static bool ContainsAtLeastWholeWord(this string[] a, string match)
-        {
-            if (a != null)
-            {
-                for (int i = 0; i < a.Length; i++)
-                    if (a[i].ContainsWholeWord(match))
-                        return true;
-                return false;
-            }
-            return false;
-        }
-        public static string Combine(this string[] s)
-        {
-            StringBuilder str = new StringBuilder();
-            for (int i = 0; i < s.Length; i++)
-                str = str.Append(s[i]);
-            return str.ToString();
-        }
-        public static string Combine(this string[] s, string separator)
-        {
-            if (s.Length > 0)
-            {
-                StringBuilder str = new StringBuilder();
-                str = str.Append(s[0]);
-                for (int i = 1; i < s.Length; i++)
-                    str = str.Append(separator + s[i]);
-                return str.ToString();
-            }
-            else return "";
-        }
-        public static string[,] Split(this string[] s, char separator)
-        {
-            List<string[]> separated = new List<string[]>();
-            for (int i = 0; i < s.Length; i++)
-                separated.Add(s[i].Split(separator));
-            return separated.ToTwoDimensionalArray();
-        }
-
-        public static string ShowValuesWithRanges(this int[] s)
-        {
-            s = s.Sort(); // Sort the values
-            string result = "";
-            if (s.Length > 0)
-            {
-                int lastShownValue = s[0];
-                int lastValueInCombo = s[0];
-                result += lastShownValue.ToString();
-                for (int i = 0; i < s.Length; i++)
-                {
-                    if (s[i] > lastValueInCombo) // Determines whether the next value is not in a row
-                    {
-                        result += (result[result.Length - 1] == '-' ? s[i - 1].ToString() : "") + ", " + s[i].ToString();
-                        lastValueInCombo = lastShownValue = s[i] + 1;
-                    }
-                    else if (s[i] == lastValueInCombo) // Determines whether the next value is in a row
-                    {
-                        if (i < s.Length - 1) // Determines whether the current index is not the last value
-                        {
-                            if (lastShownValue == lastValueInCombo) // Determines whether this is the start of a new combo
-                                result += "-";
-                        }
-                        else if (i == s.Length - 1) // Determines whether the current index is the last value
-                            result += s[i].ToString();
-                        lastValueInCombo = s[i] + 1; // Set the last index in the combo to the current index
-                    }
-                }
-            }
-            return result;
-        }
-
         public static bool Contains<T>(this T[] a, T item)
         {
             for (int i = 0; i < a.Length; i++)
@@ -410,20 +57,6 @@ namespace GDEdit.Utilities.Functions.General
                 if (l[i] == match)
                     return i;
             return -1;
-        }
-        public static List<int> BubbleSortList(this List<int> list)
-        {
-            int temp = 0;
-            List<int> l = list;
-            for (int i = 0; i < l.Count - 1; i++)
-                for (int j = 0; j < l.Count - i - 1; j++)
-                    if (l[j] > l[j + 1])
-                    {
-                        temp = l[j + 1];
-                        l[j + 1] = l[j];
-                        l[j] = temp;
-                    }
-            return l;
         }
         public static List<int> RemoveDuplicates(this List<int> l)
         {
@@ -546,140 +179,6 @@ namespace GDEdit.Utilities.Functions.General
             return lengths;
         }
 
-        public static T[] Append<T>(this T[] a, T item)
-        {
-            if (a != null)
-            {
-                T[] result = new T[a.Length + 1];
-                for (int i = 0; i < a.Length; i++)
-                    result[i] = a[i];
-                result[a.Length] = item;
-                return result;
-            }
-            else return new T[] { item };
-        }
-        public static T[] CopyArray<T>(this T[] a)
-        {
-            T[] result = new T[a.Length];
-            for (int i = 0; i < a.Length; i++)
-                result[i] = a[i];
-            return result;
-        }
-        public static T[] ExcludeFirst<T>(this T[] a, int excludedItemsCount)
-        {
-            if (excludedItemsCount <= a.Length)
-            {
-                T[] ar = new T[a.Length - excludedItemsCount];
-                for (int i = excludedItemsCount; i < a.Length; i++)
-                    ar[i - excludedItemsCount] = a[i];
-                return ar;
-            }
-            else
-                throw new ArgumentException("The excluded items must be less or equal to the size of the array.", "excludedItemsCount");
-        }
-        public static T[] ExcludeLast<T>(this T[] a, int excludedItemsCount)
-        {
-            if (excludedItemsCount <= a.Length)
-            {
-                T[] ar = new T[a.Length - excludedItemsCount];
-                for (int i = 0; i < ar.Length; i++)
-                    ar[i] = a[i];
-                return ar;
-            }
-            else
-                throw new ArgumentException("The excluded items must be less or equal to the size of the array.", "excludedItemsCount");
-        }
-        public static T[] InsertAt<T>(this T[] a, int index, T item)
-        {
-            if (a != null)
-            {
-                T[] result = new T[a.Length + 1];
-                result[index] = item;
-                for (int i = 0; i < index; i++)
-                    result[i] = a[i];
-                for (int i = index; i < a.Length; i++)
-                    result[i + 1] = a[i];
-                return result;
-            }
-            else return new T[] { item };
-        }
-        public static T[] InsertAtStart<T>(this T[] a, T item)
-        {
-            if (a != null)
-            {
-                T[] result = new T[a.Length + 1];
-                result[0] = item;
-                for (int i = 0; i < a.Length; i++)
-                    result[i + 1] = a[i];
-                return result;
-            }
-            else return new T[] { item };
-        }
-        public static T[] MoveElement<T>(this T[] a, int from, int to)
-        {
-            T[] result = new T[a.Length];
-            for (int i = 0; i < a.Length; i++)
-                result[i] = a[i];
-            result = result.RemoveAt(from);
-            result = result.InsertAt(to, a[from]);
-            return result;
-        }
-        public static T[] MoveElementToEnd<T>(this T[] a, int from)
-        {
-            T[] result = new T[a.Length];
-            for (int i = 0; i < a.Length; i++)
-                result[i] = a[i];
-            result = result.RemoveAt(from);
-            result = result.InsertAt(result.Length, a[from]);
-            return result;
-        }
-        public static T[] MoveElementToStart<T>(this T[] a, int from)
-        {
-            T[] result = new T[a.Length];
-            for (int i = 0; i < a.Length; i++)
-                result[i] = a[i];
-            result = result.RemoveAt(from);
-            result = result.InsertAt(0, a[from]);
-            return result;
-        }
-        public static T[] RemoveAt<T>(this T[] a, int index)
-        {
-            T[] result = new T[a.Length - 1];
-            for (int i = 0; i < index; i++)
-                result[i] = a[i];
-            for (int i = index + 1; i < a.Length; i++)
-                result[i - 1] = a[i];
-            return result;
-        }
-        public static T[] RemoveDuplicates<T>(this T[] a)
-        {
-            List<T> result = new List<T>();
-            for (int i = 0; i < a.Length; i++)
-                if (!result.Contains(a[i]))
-                    result.Add(a[i]);
-            return result.ToArray();
-        }
-        public static T[] Reverse<T>(this T[] a)
-        {
-            T[] result = new T[a.Length];
-            for (int i = 0; i < a.Length; i++)
-                result[a.Length - i - 1] = a[i];
-            return result;
-        }
-        public static T[] Sort<T>(this T[] ar)
-        {
-            List<T> sorted = ar.ToList();
-            sorted.Sort();
-            return sorted.ToArray();
-        }
-        public static T[] Swap<T>(this T[] ar, int a, int b)
-        {
-            T[] result = CopyArray(ar);
-            T t = ar[a];
-            ar[a] = ar[b];
-            ar[b] = t;
-            return result;
-        }
         public static List<T> InsertAtStart<T>(this List<T> a, T item)
         {
             if (a != null)
@@ -715,34 +214,6 @@ namespace GDEdit.Utilities.Functions.General
             return l;
         }
 
-        public static List<T> Clone<T>(this List<T> l)
-        {
-            List<T> result = new List<T>();
-            for (int i = 0; i < l.Count; i++)
-                result.Add(l[i]);
-            return result;
-        }
-        public static List<List<T>> Clone<T>(this List<List<T>> l)
-        {
-            List<List<T>> result = new List<List<T>>();
-            for (int i = 0; i < l.Count; i++)
-            {
-                result.Add(new List<T>());
-                for (int j = 0; j < l[i].Count; j++)
-                    result[i].Add(l[i][j]);
-            }
-            return result;
-        }
-        public static bool ContainsAll<T>(this List<T> list, List<T> containedList)
-        {
-            if (containedList.Count != list.Count)
-                return false;
-            List<T> tempList = list.Clone();
-            List<T> tempContained = containedList.Clone();
-            for (int i = 0; i < tempContained.Count; i++)
-                tempList.Remove(tempContained[i]);
-            return tempList.Count == 0;
-        }
         public static bool ContainsOrdered(this List<bool> list, List<bool> containedList)
         {
             for (int i = 0; i < list.Count - containedList.Count; i++)
@@ -776,17 +247,6 @@ namespace GDEdit.Utilities.Functions.General
             }
             return false;
         }
-        // Probably O(n) complexity? Or is it O(n^2) because it's looping on an O(n) complexity function? I don't know how this works, someone teach me
-        // This is probably O(n^2) complexity, my past self
-        public static bool ContainsUnordered<T>(this List<T> list, List<T> containedList)
-        {
-            List<T> tempList = containedList.Clone();
-            List<T> tempContained = containedList.Clone();
-            for (int i = 0; i < tempList.Count; i++)
-                tempList.Remove(tempContained[i]);
-            return list.Count - containedList.Count == tempList.Count;
-        }
-
         public static TValue ValueOrDefault<TKey, TValue>(this Dictionary<TKey, TValue> d, TKey key) => d.ContainsKey(key) ? d[key] : default;
 
         public static int Max(this int[] numbers)
@@ -992,14 +452,6 @@ namespace GDEdit.Utilities.Functions.General
             for (int i = 0; i < a.Length; i++)
                 result[i] = a[i].ToString();
             return result;
-        }
-        public static T[,] ToTwoDimensionalArray<T>(this List<T[]> l)
-        {
-            T[,] ar = new T[l.Count, l.GetLengths().Max()];
-            for (int i = 0; i < l.Count; i++)
-                for (int j = 0; j < l[i].Length; j++)
-                    ar[i, j] = l[i][j];
-            return ar;
         }
         public static List<List<T>> ToList<T>(this T[,] ar)
         {
